Integration Test; JPA working

+++ jpa/src/main/java/io/freedriver/autonomy/jpa/converter/measurement/TemporalMeasurementConverter.java
@@ -7,15 +7,14 @@ import java.math.BigDecimal;
 
 public abstract class TemporalMeasurementConverter<TM extends TemporalMeasurement<TM>> extends MeasurementConverter<TM> {
     @Override
-    public Double convertToDatabaseColumn(TM tm) {
+    public BigDecimal convertToDatabaseColumn(TM tm) {
         return tm.scaleTo(UnitPrefix.ONE).getValue()
-                .getValue()
-                .doubleValue();
+                .getValue();
     }
 
     @Override
-    public TM convertToEntityAttribute(Double value) {
-        return construct(BigDecimal.valueOf(value), UnitPrefix.ONE)
+    public TM convertToEntityAttribute(BigDecimal value) {
+        return construct(value, UnitPrefix.ONE)
                 .normalize();
     }
